test: move ipcRenderer spec to main runner (#20093)
